Fixed various bugs as in attached description.
Fixed a bug with alpha boost that meant only Boost_O could ever work.  Simplified mode switching to a maximum of two shader passes.  Reduced the number of samplers required from seven to four.
Corrected an implicit cast of float2 to float in fmod(), which broke the outgoing and dissolve transitions in Linux and Mac.